Implementation of new features, and a lot of bug corrections.
Now it is possible to send signals to the jobs in the queue and the
ones which are already running.
The possible signals are: kill, pause, continue and queue. The last
means the job will stop running in the current host and will be brought
back to the queue, with the results computed so far saved.
It is also possible to set a list of hosts where the job can be executed
and also change the priority of the job.

Another bug free (for now) feature is the setting and visualization of the
calendar of the hosts.