Important changes to the structure of the JobManager
Now the major part of the communications between clients and server carries
only views of the jobs. A jobview is a class with most of the attributes of
the job but the input_files, output_files and execution_script. Sending
only a jobview for communication reduces the load on the network, since the
three missing attributes are responsible for most of the size of the object.

Changes suggested by FAC in the last meeting:
    Specification of initial and final time in pyjob_configs_set, instead of
        hour and minute options
    Correction of bug in pyjob_qsub. Now it is possible to submit jobs
        without the possiblehosts setting
    Change o qstat to show the creation time and time running informations.
    Now it is possible to choose which job information to show.